add view attendees, bug fixed, delete invite

+++ modules/domain/gateway/src/main/java/com/doneit/ascend/domain/gateway/common/mapper/to_entity/UserMapper.kt
@@ -115,7 +115,7 @@ fun RateResponse.toEntity(): RateEntity {
 
 fun SearchUsersResponse.toEntity(): AttendeeEntity{
     return AttendeeEntity(
-        id.toInt(),
+        id,
         fullName,
         email,
         image?.url
